+++ internal/controller/flow_falco_test.go
@@ -1,6 +1,7 @@
 package controller
 
 import (
+	pb "github.com/illumio/cloud-operator/api/illumio/cloud/k8sclustersync/v1"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -92,6 +93,142 @@ func (suite *ControllerTestSuite) TestFilterIllumioTraffic() {
 	for name, tt := range tests {
 		suite.Run(name, func() {
 			result := filterIllumioTraffic(tt.input)
+			assert.Equal(suite.T(), tt.expected, result)
+		})
+	}
+}
+
+func (suite *ControllerTestSuite) TestCreateLayer4Message() {
+	tests := map[string]struct {
+		proto          string
+		srcPort        uint32
+		dstPort        uint32
+		ipVersion      string
+		expected       *pb.Layer4
+		expectedErrMsg string
+	}{
+		"TCP protocol": {
+			proto:     "tcp",
+			srcPort:   80,
+			dstPort:   8080,
+			ipVersion: "",
+			expected: &pb.Layer4{
+				Protocol: &pb.Layer4_Tcp{
+					Tcp: &pb.TCP{
+						SourcePort:      80,
+						DestinationPort: 8080,
+						Flags:           &pb.TCPFlags{},
+					},
+				},
+			},
+			expectedErrMsg: "",
+		},
+		"UDP protocol": {
+			proto:     "udp",
+			srcPort:   123,
+			dstPort:   456,
+			ipVersion: "",
+			expected: &pb.Layer4{
+				Protocol: &pb.Layer4_Udp{
+					Udp: &pb.UDP{
+						SourcePort:      123,
+						DestinationPort: 456,
+					},
+				},
+			},
+			expectedErrMsg: "",
+		},
+		"ICMP protocol with IPv4": {
+			proto:     "icmp",
+			srcPort:   0,
+			dstPort:   0,
+			ipVersion: "ipv4",
+			expected: &pb.Layer4{
+				Protocol: &pb.Layer4_Icmpv4{
+					Icmpv4: &pb.ICMPv4{},
+				},
+			},
+			expectedErrMsg: "",
+		},
+		"ICMP protocol with IPv6": {
+			proto:     "icmp",
+			srcPort:   0,
+			dstPort:   0,
+			ipVersion: "ipv6",
+			expected: &pb.Layer4{
+				Protocol: &pb.Layer4_Icmpv6{
+					Icmpv6: &pb.ICMPv6{},
+				},
+			},
+			expectedErrMsg: "",
+		},
+		"Unknown protocol": {
+			proto:          "unknown",
+			srcPort:        0,
+			dstPort:        0,
+			ipVersion:      "",
+			expected:       nil,
+			expectedErrMsg: "unknown protocol: unknown",
+		},
+	}
+
+	for name, tt := range tests {
+		suite.Run(name, func() {
+			result, err := CreateLayer4Message(tt.proto, tt.srcPort, tt.dstPort, tt.ipVersion)
+			if tt.expectedErrMsg != "" {
+				assert.Error(suite.T(), err)
+				assert.EqualError(suite.T(), err, tt.expectedErrMsg)
+			} else {
+				assert.NoError(suite.T(), err)
+				assert.Equal(suite.T(), tt.expected, result)
+			}
+		})
+	}
+}
+
+func (suite *ControllerTestSuite) TestCreateLayer3Message() {
+	tests := map[string]struct {
+		source      string
+		destination string
+		ipVersion   string
+		expected    *pb.IP
+	}{
+		"IPv4": {
+			source:      "192.168.0.1",
+			destination: "192.168.0.2",
+			ipVersion:   "ipv4",
+			expected: &pb.IP{
+				Source:      "192.168.0.1",
+				Destination: "192.168.0.2",
+				IpVersion:   pb.IPVersion_IP_VERSION_IPV4,
+			},
+		},
+		"IPv6": {
+			source:      "fe80::1",
+			destination: "fe80::2",
+			ipVersion:   "ipv6",
+			expected: &pb.IP{
+				Source:      "fe80::1",
+				Destination: "fe80::2",
+				IpVersion:   pb.IPVersion_IP_VERSION_IPV6,
+			},
+		},
+		"Unspecified IP version": {
+			source:      "192.168.0.1",
+			destination: "192.168.0.2",
+			ipVersion:   "unknown",
+			expected: &pb.IP{
+				Source:      "192.168.0.1",
+				Destination: "192.168.0.2",
+				IpVersion:   pb.IPVersion_IP_VERSION_IP_NOT_USED_UNSPECIFIED,
+			},
+		},
+	}
+
+	for name, tt := range tests {
+		suite.Run(name, func() {
+			result, err := createLayer3Message(tt.source, tt.destination, tt.ipVersion)
+			assert.NoError(suite.T(), err)
 			assert.Equal(suite.T(), tt.expected, result)
 		})
 	}
